% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabICL-predict.R
\name{predict.tab_icl}
\alias{predict.tab_icl}
\alias{augment.tab_icl}
\title{Predict using TabICL}
\usage{
\method{predict}{tab_icl}(object, new_data, type = NULL, ...)

\method{augment}{tab_icl}(x, new_data, type = NULL, ...)
}
\arguments{
\item{object, x}{A \code{tab_icl} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{For classification, one of \code{"class"} or \code{"prob"}. Defaults to
all available prediction columns.}

\item{...}{Not used.}
}
\value{
A tibble of predictions.
}
\description{
Predict using TabICL
}
