% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabICL-fit.R
\name{tab_icl}
\alias{tab_icl}
\alias{tab_icl.default}
\alias{tab_icl.data.frame}
\alias{tab_icl.matrix}
\alias{tab_icl.formula}
\alias{tab_icl.recipe}
\title{Fit a TabICLv2 model}
\usage{
tab_icl(x, ...)

\method{tab_icl}{default}(x, ...)

\method{tab_icl}{data.frame}(x, y, training_set_limit = Inf, control = control_tab_icl(), ...)

\method{tab_icl}{matrix}(x, y, training_set_limit = Inf, control = control_tab_icl(), ...)

\method{tab_icl}{formula}(
  formula,
  data,
  training_set_limit = Inf,
  control = control_tab_icl(),
  ...
)

\method{tab_icl}{recipe}(x, data, training_set_limit = Inf, control = control_tab_icl(), ...)
}
\arguments{
\item{x}{A data frame, matrix, recipe, or formula.}

\item{...}{Not currently used.}

\item{y}{Outcome vector for the data-frame and matrix interfaces.}

\item{training_set_limit}{Maximum number of training rows retained before
fitting. Use \code{Inf} to disable downsampling.}

\item{control}{A list from \code{\link[=control_tab_icl]{control_tab_icl()}}.}

\item{formula}{Formula for the formula interface.}

\item{data}{Data frame for formula and recipe interfaces.}
}
\value{
A \code{tab_icl} model object.
}
\description{
\code{tab_icl()} fits the Python \code{tabicl} classifier or regressor through
\code{reticulate} and returns an R object with \code{predict()} and \code{augment()}
methods.
}
